# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Swagger for Jobbi Back-end\n\nSchemes: [http, https]",
    "title": "Jobbi Back-end",
    "contact": {
      "name": "Maxime Denuit",
      "email": "maxime@nightborn.be"
    },
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/searches/{searchId}": {
      "get": {
        "description": "type id struct\nGet search by id.\nReturn search",
        "produces": [
          "application/json"
        ],
        "tags": [
          "searches"
        ],
        "operationId": "GetSearchById",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "type id struct\nCreate search.\nReturn search",
        "produces": [
          "application/json"
        ],
        "tags": [
          "searches"
        ],
        "operationId": "ModifySearch",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "description": "search",
            "name": "search",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/SearchDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/searches/{searchId}/companies": {
      "get": {
        "description": "type id struct\nGet Companies by searchId.\nReturn company",
        "produces": [
          "application/json"
        ],
        "tags": [
          "companies"
        ],
        "operationId": "GetCompaniesBySearchId",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "All the companies",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CompanyDTO"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "type id struct\nCreate company.\nReturn company",
        "produces": [
          "application/json"
        ],
        "tags": [
          "companies"
        ],
        "operationId": "AddCompany",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "description": "company",
            "name": "company",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CompanyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CompanyDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/searches/{searchId}/companies/{companyId}": {
      "get": {
        "description": "type id struct\nGet company by id.\nReturn company",
        "produces": [
          "application/json"
        ],
        "tags": [
          "companies"
        ],
        "operationId": "GetCompanyById",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "companyId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CompanyDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "type id struct\nModify company.\nReturn company",
        "produces": [
          "application/json"
        ],
        "tags": [
          "companies"
        ],
        "operationId": "ModifyCompany",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "companyId",
            "in": "path",
            "required": true
          },
          {
            "description": "company",
            "name": "company",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/CompanyDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CompanyDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "type id struct\nDelete company.\nReturn true or error",
        "produces": [
          "application/json"
        ],
        "tags": [
          "companies"
        ],
        "operationId": "DeleteCompany",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "companyId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/searches/{searchId}/networks": {
      "get": {
        "description": "type id struct\nGet networks by searchId.\nReturn network",
        "produces": [
          "application/json"
        ],
        "tags": [
          "networks"
        ],
        "operationId": "GetNetworksBySearchId",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/NetworkDTO"
              }
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "description": "type id struct\nPost  network.\nReturn  network",
        "produces": [
          "application/json"
        ],
        "tags": [
          "networks"
        ],
        "operationId": "AddNetwork",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "description": "test",
            "name": "network",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NetworkDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/NetworkDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/searches/{searchId}/networks/{networkId}": {
      "get": {
        "description": "type id struct\nGet  network by id.\nReturn  network",
        "produces": [
          "application/json"
        ],
        "tags": [
          "networks"
        ],
        "operationId": "GetNetworkById",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "networkId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/NetworkDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "description": "type id struct\nPut  network.\nReturn  network",
        "produces": [
          "application/json"
        ],
        "tags": [
          "networks"
        ],
        "operationId": "ModifyNetwork",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "networkId",
            "in": "path",
            "required": true
          },
          {
            "description": "test",
            "name": "network",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NetworkDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/NetworkDTO"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "description": "type id struct\nDelete network.\nReturn true or error",
        "produces": [
          "application/json"
        ],
        "tags": [
          "networks"
        ],
        "operationId": "DeleteNetwork",
        "parameters": [
          {
            "type": "string",
            "description": "test",
            "name": "searchId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "test",
            "name": "networkId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "boolean"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "CompanyDTO": {
      "description": "An company",
      "type": "object",
      "properties": {
        "description": {
          "description": "The description",
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "description": "The id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Id"
        },
        "link": {
          "description": "The link",
          "type": "string",
          "x-go-name": "Link"
        },
        "searchId": {
          "description": "The SearchID",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "SearchID"
        },
        "title": {
          "description": "The title",
          "type": "string",
          "x-go-name": "Title"
        },
        "userId": {
          "description": "The UserID",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/philvc/jobbi-api/contract"
    },
...
    "NetworkDTO": {
      "description": "An network",
      "type": "object",
      "properties": {
        "description": {
          "description": "The description",
          "type": "string",
          "x-go-name": "Description"
        },
        "email": {
          "description": "The email",
          "type": "string",
          "x-go-name": "Email"
        },
        "firstName": {
          "description": "The firstName",
          "type": "string",
          "x-go-name": "FirstName"
        },
        "id": {
          "description": "The id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Id"
        },
        "lastName": {
          "description": "The lastName",
          "type": "string",
          "x-go-name": "LastName"
        },
        "link": {
          "description": "The link",
          "type": "string",
          "x-go-name": "Link"
        },
        "phoneNumber": {
          "description": "The phoneNumber",
          "type": "string",
          "x-go-name": "PhoneNumber"
        },
        "searchId": {
          "description": "The search id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "SearchID"
        },
        "userId": {
          "description": "The user id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/philvc/jobbi-api/contract"
    },
...
    "SearchDTO": {
      "description": "An search",
      "type": "object",
      "required": [
        "id",
        "description",
        "title"
      ],
      "properties": {
        "description": {
          "description": "The description",
          "type": "string",
          "x-go-name": "Description"
        },
        "id": {
          "description": "The id",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Id"
        },
        "title": {
          "description": "The title",
          "type": "string",
          "x-go-name": "Title"
        },
        "userId": {
          "description": "UserId",
          "type": "integer",
          "format": "uint64",
          "x-go-name": "UserID"
        }
      },
      "x-go-package": "github.com/philvc/jobbi-api/contract"
    },
...
  }
}